Rename modules: lower case, remove +
