+++ networkx/algorithms/isomorphism/vf2pp_helpers/candidates.py
