+++ networkx/algorithms/isomorphism/vf2pp_helpers/feasibility.py
