+++ networkx/algorithms/isomorphism/vf2pp_helpers/node_ordering.py
